requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

pyspark==4.1.1
py4j==0.10.9.9

pandas==2.2.2
numpy==1.26.4

kafka-python==2.0.2
requests==2.31.0

mlflow==2.12.1

fastapi==0.110.0
uvicorn==0.29.0

pyspark==4.1.1
py4j==0.10.9.9

pandas==2.2.2
numpy==1.26.4

kafka-python==2.0.2
requests==2.31.0

mlflow==2.12.1